create article component
